fix: don't add additional blank audio track (see details)
When we create a new timeline a 2.0 Audio track already exists, this
could potentially be a different SubType than our source material's
first track.

Alternatively we could create a new video track and start at
TrackIndex + 1, but that leaves us with the empty default tracks that
currently I can't remove via the API.

It's more of a hassle to remove the empty tracks than the likelihood
that the 1st audio track isn't a stereo 2.0 track, so for now this is
fine.