feat: use FFI Windows API to avoid launching terminals in the foreground
There has got to be a better solution than this, but it will work for
now.

We're basically digging into the windows C api just so we don't have to
launch terminal windows every time we run a ffmpeg command which seems
absurd to me, but FFI is built into the lua with Resolve and it seems
to be functional.